+++ src/app/onboarding/_components/onBoardingStep/index.tsx
@@ -1,8 +1,9 @@
 import { useRouter } from "next/navigation";
-import { FC, ReactElement, useEffect, useState } from "react";
+import { FC, ReactElement, useEffect } from "react";
 import { useFormContext } from "react-hook-form";
 
 import { ProgressBar } from "@/components/core/Progress";
+import { useOnBoardingStep } from "@/hooks/useOnBoardingStep";
 import { DefaultHeader } from "@/layout/Mobile/MobileHeader";
 import {
   FestivalCategory,
@@ -13,7 +14,6 @@ import {
 } from "@/model/onboarding";
 import { delay } from "@/utils/delay";
 
-import { ONBOARDING } from "../../_constants";
 import OnBoardingButton from "./OnBoardingButton";
 import OnBoardingCategories from "./OnBoardingCategories";
 import OnBoardingCompanies from "./OnBoardingCompanies";
@@ -35,8 +35,8 @@ const OnBoardingContainer: FC<Props> = ({
 }) => {
   const router = useRouter();
 
-  const { INITIAL_STEP, TOTAL_STEP } = ONBOARDING;
-  const [currentStep, setCurrentStep] = useState<number>(INITIAL_STEP);
+  const { currentStep, handleNextStep, handlePrevStep, ONBOARDING } =
+    useOnBoardingStep();
   const { trigger, handleSubmit } = useFormContext<OnboardingModel>();
 
   // * 첫 렌더링에 유효성 검사 1회 실행
@@ -44,20 +44,10 @@ const OnBoardingContainer: FC<Props> = ({
     trigger();
   }, []);
 
-  const handleStepChange = (stepChange: number) => {
-    setCurrentStep((prev) => {
-      const newStep = prev + stepChange;
-      return newStep >= INITIAL_STEP && newStep <= TOTAL_STEP ? newStep : prev;
-    });
-  };
-
-  const handlePrevStep = () => handleStepChange(-1);
-  const handleNextStep = () => handleStepChange(1);
-
   const onSubmit = async (data: OnboardingModel) => {
     await delay(5000);
     console.log("🚀 ~ onSubmit ~ data:", data);
-    await router.replace("/onboarding/complete");
+    router.replace("/onboarding/complete");
   };
 
   const renderCurrentStep: {
